refactor: externalize create_csv()
put create_csv() in csv_management/csv_creator.py and import it in main program
